+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * [BUGFIX] Work around issue preventing feature execution on Ruby 3.5.0dev (by [@faisal][])
 * [CHANGE] Add changes or suppress warnings for issues found by newer rubocop (by [@faisal][])
 * [CHANGE] Update CI checkout action to v4 (by [@faisal][])
+* [CHANGE] Run RubyCritic outside of the project without losing the churn value (by [@juanvqz][])
 
 # v4.9.2 / 2025-04-08 [(commits)](https://github.com/whitesmith/rubycritic/compare/v4.9.1...v4.9.2)